Added alphabetic sorted zones in each countries in GeoZones page
Make scripts that check sorting and geo zones and countries in Admin
litecart.
1) On the Country page:
a) verify that countries are in alphabetical order;
b) for those countries in which the number of zones is different from
zero:  open the page of the country and there to check, that the zones
are arranged in alphabetical order;
2) On the Geo Zones page go into each country and to check that the
zones are arranged alphabetically.